# specs/001-core-ventas-inventario/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Comercio Inteligente - Core API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
  /dashboard/summary:
    get: {summary: Consultar indicadores y ventas recientes del resumen, responses: {'200': {description: Resumen operativo}}}
  /categories:
    get: {summary: Listar categorías, responses: {'200': {description: Lista}}}
    post: {summary: Crear categoría, responses: {'201': {description: Creada}, '409': {description: Código duplicado}}}
  /suppliers:
    get: {summary: Listar proveedores, responses: {'200': {description: Lista}}}
    post: {summary: Crear proveedor, responses: {'201': {description: Creado}, '409': {description: Código duplicado}}}
  /locations:
    get: {summary: Listar sucursales y almacenes, responses: {'200': {description: Lista}}}
    post: {summary: Crear ubicación, responses: {'201': {description: Creada}, '409': {description: Código duplicado}}}
  /products:
    get:
      summary: Buscar productos vigentes
      parameters:
        - { name: q, in: query, schema: { type: string } }
      responses:
        '200': { description: Productos con precio y disponibilidad }
    post:
      summary: Crear producto
      responses:
        '201': { description: Producto creado }
  /inventory:
    get:
      summary: Consultar existencias, lotes y alertas
      responses:
        '200': { description: Inventario consultado }
  /inventory/movements:
    get:
      summary: Consultar movimientos trazables de inventario
      parameters:
        - { name: product_id, in: query, schema: { type: string } }
        - { name: location_id, in: query, schema: { type: string } }
        - { name: type, in: query, schema: { type: string } }
        - { name: date_from, in: query, schema: { type: string, format: date } }
        - { name: date_to, in: query, schema: { type: string, format: date } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 1000 } }
      responses:
        '200': { description: Movimientos con producto, origen, ubicación, costo, fecha y responsable }
        '422': { description: Filtro inválido }
  /inventory/receipts:
    post:
      summary: Registrar entrada directa de inventario
      parameters:
        - { name: Idempotency-Key, in: header, required: true, schema: { type: string } }
      responses: {'201': {description: Entrada registrada}}
  /sales:
    post:
      summary: Crear borrador de venta
      responses:
        '201': { description: Venta creada }
  /sales/{saleId}/confirm:
    post:
      summary: Confirmar venta de forma atómica
      parameters:
        - { name: saleId, in: path, required: true, schema: { type: string } }
        - { name: Idempotency-Key, in: header, required: true, schema: { type: string } }
      responses:
        '200': { description: Venta confirmada }
        '409': { description: Stock o versión en conflicto }
  /purchase-orders:
    get:
      summary: Listar órdenes de compra
      responses: {'200': {description: Lista}}
    post:
      summary: Crear orden de compra
      responses:
        '201': { description: Orden creada }
  /purchase-orders/{orderId}/receipts:
    post:
      summary: Registrar recepción total o parcial
      parameters:
        - { name: orderId, in: path, required: true, schema: { type: string } }
        - { name: Idempotency-Key, in: header, required: true, schema: { type: string } }
      responses:
        '201': { description: Recepción registrada }
  /stock-counts:
    post:
      summary: Abrir conteo de inventario
      responses:
        '201': { description: Conteo abierto }
  /stock-counts/{countId}/approve:
    post:
      summary: Aprobar diferencias y crear ajustes
      parameters:
        - { name: countId, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Ajustes registrados }
  /lost-sales:
    post:
      summary: Registrar demanda no atendida
      responses:
        '201': { description: Venta perdida registrada }
  /returns:
    post:
      summary: Registrar devolución e inspección
      responses:
        '201': { description: Devolución registrada }